# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Simple calendar for booking events
  description: Simple calendar for booking events
  contact:
    email: retfalvi.levente@virgo.hu
  version: 0.0.1-SNAPSHOT

tags:
  - name: calendar
    description: Managing calendar events.

paths:
  /calendar/events/at/{dateTime}/organizer:
    get:
      tags:
        - calendar
      summary: Get an event by the given date and time
      description: Find the ongoing event with the given date and time. Also works with past events
      operationId: organizerOfEventByDateTime
      parameters:
        - name: dateTime
          in: path
          required: true
          description: Date and time for the query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Event successfully returned
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Event not found

  /calendar/events/:
    post:
      tags:
        - calendar
      summary: Create an event in the calendar
      description: Create an event in the calendar
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '200':
          description: Event successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid request, failed to create event
  
  /calendar/week/{date}/event:
    get:
      tags:
        - calendar
      summary: Obtain the schedule for the given week
      description: Obtain the schedule for the given week, the start of the week is preconfiguration
      operationId: eventsOfWeek
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successfully returned the scheduled events for the given week
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'

  /calendar/week/{date}/timeslots/:
    get:
      tags:
        - calendar
      summary: Obtain the available time slots for the given week
      description: Obtain the available time slots for the given week, the start of the week is preconfiguration
      operationId: availableTimeSlotsOfWeek
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successfully returned the available time slots for the given week
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'

components:
  schemas:
    Event:
      type: object
      required:
        - startTime
        - endTime
        - organizer
      properties:
        id:
          type: integer
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        organizer:
          type: string
    TimeSlot:
      type: object
      required:
        - start
        - end
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time